reorganize filter code
